# Add support for RegExp patterns with named capture groups