--- a/package.json
+++ b/package.json
@@ -29,7 +29,7 @@
     "routing"
   ],
   "dependencies": {
-    "regexparam": "^1.2.0"
+    "regexparam": "^1.3.0"
   },
   "devDependencies": {
     "bundt": "0.4.0",