The support of RegExp patterns with named capture groups will significantly increase flexiblity of Trouter – it will allow to express any complex patterns for the cases where the regexparam patterns are not enough.

I’ve run bench on Node v10.14.2 and this change doesn’t have any provable negative effect on the performance.

It will increase size of index.mjs by ~11 %.

```diff
 Filename         Filesize  (gzip)
-index.js          1.18 kB  519 B 
+index.js          1.33 kB  579 B
-index.mjs         1.17 kB  518 B
+index.mjs         1.32 kB  577 B
```