@lukeed 2019-07-04
Cool! I'm traveling today but will do a careful comb through it later -- performance testing too (I have some local suites) since Trouter is the foundation of Polka

Thanks