@lukeed 2019-07-04
Okay, this looks good. I haven't done any performance testing, but I will resume at a later date.

I think I'll make changes to `regexparam` so that the changes here are as bit simpler, and it will allow `navaid` to _also_ allow custom, named RegExps for client-side routing. Please keep this PR open :)

Feeling a bit under the weather right now (probably due to traveling) so I'll have to resume another day... likely this weekend.

Thanks again~!