@lukeed 2019-07-09
Final size is `558B` and unfortunately there _is_ measurable performance impact, but it's been mitigated to the best of my (current) ability