@paulocoghi 2019-07-09
> Final size is 558B and unfortunately there is measurable performance impact, but it's been mitigated to the best of my (current) ability

This is exactly my concern: performance reduction when we _don't_ use RegExp.